turn Header into a decorator class so instances may be mapped to extensions of the base RegularHeader class when different parsing strategies are required for a particular header
